server/requirements.txt
fastapi==0.135.1
numpy==2.4.2
protobuf==5.29.6
pydantic==2.12.5
pydantic_settings==2.13.1
python-dotenv==1.2.2
tavily-python==0.7.22
trafilatura==2.0.0
uvicorn==0.41.0
websockets==16.0
google-genai==1.2.0
lxml_html_clean==0.4.4
